Move metric functions to its own module
- Renamed functions to better reflect their purpose
